requirements.txt
anthropic>=0.25.0
openai>=1.30.0
pydantic>=2.0.0
PyYAML>=6.0
requests>=2.31.0
